requirements.txt
streamlit
pdfplumber
easyocr
pillow
requests
python-dotenv
torch
langchain
langchain-community
